requirements.txt
ccxt
pandas
matplotlib
mplfinance
ta
python-telegram-bot==13.15
openpyxl
